+++ WooCommerce/src/main/kotlin/com/woocommerce/android/ui/orders/wooshippinglabels/packages/datasource/WooShippingLabelPackageRepository.kt
@@ -21,24 +21,4 @@ class WooShippingLabelPackageRepository @Inject constructor(
             ?.let { WooResult(it) }
             ?: WooResult(error)
     }
-
-    data class StorePackagesDAO(
-        val savedPackages: List<PackageDAO>,
-        val carrierPackages: List<CarrierDAO>
-    )
-
-    data class PackageDAO(
-        val id: String,
-        val name: String,
-        val dimensions: String
-    )
-
-    data class CarrierDAO(
-        val packageGroup: List<CarrierPackageGroupDAO>
-    )
-
-    data class CarrierPackageGroupDAO(
-        val description: String,
-        val packages: List<PackageDAO>
-    )
 }
